Deployment settings (#122)
This PR adds the Pulumi Deployment Settings resource to the provider. The serde code in provider/deployment_settings.go is somewhat unpleasant to behold but I have tested it via automated tests and lots of local testing (especially for the import case). Definitely open to it if there are any quick wins to improve some of this code, but I'd prefer to avoid too much refactoring at the moment and would prefer to get it in the hands of users.

In the future we should definitely consider migrating to the go framework as it will surely cut down much of the boilerplate associated with a hand-rolled provider.

Fixes: #97