Remove the need for cmd/pulumi-gen-pulumiservice (#240)
We don't need to maintain our own gen code, we can use the pulumi binary
for that. This makes it important that we use the right version of the
pulumi binary, so we use the Makefile to ensure the version of pulumi
used matches the version of pulumi we depend on in `provider/go.mod`.

We use a similar pattern in other native providers.

I'm tagging @pierskarsenbarg and @komalali since they were the suggested
reviewers. If I should tag someone else here, please let me know.