Fixing broken tests (#326)
After an update to pulumi CLI, tests started failing. Most of the issues
were on refresh, and refresh used to be skipped for most tests, which
made it more confusing why the same test failed in yaml, but succeeded
in other languages.

Summary
- Removed SkipRefresh flags everywhere except Deployment Settings,
because this is just a bandaid to make broken resources test pass
(Deployment Settings are currently broken, with fix PR blocked by these
tests failing)
- Reverted `req.GetOldInputs()` back into `req.GetOlds()`, which is not
a good change, but the proper way would be to overhaul how AccessTokens
work (currently Diff uses internal `__inputs` property which should no
longer be used)
- Adjusted go and python environment yaml string to make refresh work
(it's quite ugly, but we do have an issue to add auto-formatting)
- Fixed Schedules and Environment Read() outputting properties instead
of inputs, causing noisy refresh diff due to discrepancy in input and
output schema (this used to be silent, became noisy after a pulumi
engine update)
- Added logic to set `displayName` of Team resource input to name if not
supplied. This is what Pulumi Service does anyway, which caused a noisy
diff