Stack Resource (#357)
Create a Stack resource, so users can manage stacks within their pulumi
program. This can enable, for example, the ability to ascribe naming
conventions to stacks, or to create a stack with deployment settings and
a schedule to regularly run updates, etc.

Tested this by updating one of the existing deployment settings tests to
use this resource as the underlying stack. Also tested `import` and
`refresh` manually.

Fixes #171