replaceOnChanges -> willReplaceOnChanges
replaceOnChanges causes the generated sdks to apply an extra replaceOnChanges resource option that it passes through to the provider plugin.  However, the yaml language doesn't have a generated sdk, so it doesn't work properly there.

Instead, we now check for replacement in the provider plugin's Diff method instead, so it always works, and document this with `willReplaceOnChanges` (which is only used for documentation).